feat(Config): Setup ApplicationConfig
The userDetailsService() method fetches user details from a custom
implementation of UserRepository based on the email of the user.

The
authenticationProvider() method creates an instance of
DaoAuthenticationProvider, which uses the userDetailsService() method
and passwordEncoder() method to fetch user details and encode the
password.

The authenticationManager() method creates an instance of
AuthenticationManager which authenticates the user based on their
username and password.

Finally, the passwordEncoder() method creates an
instance of BCryptPasswordEncoder, which is used to encode the password.