feat(SecurityConfiguration): Setup FilterChain
The configuration disables Cross-Site Request Forgery (CSRF) protection
and specifies that authentication is required for all paths except for
"/api/v1/auth/**". It also sets the session creation policy to STATELESS
and adds a JwtAuthenticationFilter before the
UsernamePasswordAuthenticationFilter. Additionally, it sets up a logout
URL and a logout handler.

This commit also includes the required
dependencies for this configuration, which include
JwtAuthenticationFilter and AuthenticationProvider.