feat(LogoutService): implement logout logic
he LogoutService class implements LogoutHandler to handle user logout.

When logout method is called,
it checks the presence of an
"Authorization" header in the HTTP request and if it contains a JWT
token,
it retrieves the token from the token repository and marks it as
expired and revoked.
The SecurityContextHolder is also cleared to
remove any associated security context.