feat(Entity): Setup VoteRecordEntity
Entity corresponding fields and annotations for persistence in the
database.
The entity includes fields such as id, userId, username, and
voteItemId.

The @Entity annotation is used to mark the class as an
entity, while the @Table annotation specifies the name of the database
table to map the entity.

The fields are annotated with @Column to map
them to the corresponding columns in the database.

The entity is also
annotated with @Builder, @NoArgsConstructor, @AllArgsConstructor,
@Getter, and @Setter annotations for convenience.