feat(Entity): Setup VoteItemEntity
This Entity represents the items that can be voted on.
The entity has
two properties:

itemId,which is an auto-generated unique
identifier

itemName, which is a non-null String representing the name
of the item

The class is annotated with the necessary JPA annotations
to map it to a database table named "vote_items".