feat(Controller): Implement Vote Controller
VoteController is responsible for handling voting requests
and
responses.

The class has two methods:
getVoteItemResults()
retrieves a
list of vote item results and returns them to the
client
voteRecords()
allows a user to vote on a set of selected vote
items.

If the vote is
successful, the method returns an HTTP 200
response with a "Success!"
message. Otherwise, it returns an HTTP 400
response with a "Bad
Request!" message.

The user's ID is extracted
from the authentication
token and passed to the voteRecordService along
with the selected vote
items.