feat(RegisterView): Setup btn and list
Register provides a form for users to register an account.
The form
contains input fields for the user's username, email, and password.
The
form submission triggers the register() function,
which sends a POST
request to the backend server to register the user.
If the registration
is successful, the user's access token and refresh token are saved to
the store using Vuex,
and the user is redirected to the homepage.

This
component uses Vue 3 and relies on Vuex, Vue Router, and Axios. The
program uses RESTful API to communicate with the backend server.